proprification du code et suppression de warnings

+++ src/libwatermark/watermark/LSBEncode.h
@@ -21,9 +21,8 @@ class LSBEncode : public WatermarkBase<data_type>
 			// Recopier cette ligne
 			auto& channelsData = static_cast<CData<data_type>*>(data.get())->_data;
 			short bit16 = 0x0001;
-			short nonbit16 = ~bit16;
+			short nonbit16 = (short) ~bit16;
 
-			// Petit exemple qui va multiplier tout par _gain.
 			for(int j = 0; j < channelsData.size(); j++)
 			{
 				auto& sampleData = channelsData[j];
